# docs/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Ekklesia Voting Backend API — v1
  description: |
    Hydra-backed live surface for the Ekklesia voting platform. v1 is paired
    with the v0 archival surface (see `openapi.yaml`) — the two coexist.

    ## What lives here

    - **Unified ballot listing** across legacy + Hydra sources.
    - **Broker endpoints** for Hydra votes: draft the signing payload,
      collect signatures (key-based or m-of-n multisig), submit to Hydra,
      surface confirmation artifacts.
    - **Admin lifecycle**: `/prepare`, `/start`, `/finalize`, the stepped
      close sequence (`/settle/burn` → `/settle/finalize` → `/settle/close`),
      plus inspection helpers (`/count`, `/queue/*`, `/head-info`,
      `/results/recover`) and `/voting-power` + `/certify` — role-gated.
    - **Public integrator surface** under `/public/*` — API-key gated,
      rate-limited, read-only.

    ## Authentication

    - `verifyToken` (JWT cookie or Authorization header) for voter session
      endpoints.
    - Same JWT + admin allowlist/claim for `/admin/*`.
    - API key (Bearer or x-api-key) for `/public/*`.
  version: 1.0.0
  contact:
    name: Adam Dean, Mad Orkestra
    license: ISC
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://docs.ekklesia.vote/api
    description: Documentation reference host
...
paths:
...
  /results/ballot/{ballotId}:
    get:
      tags: [Results]
      summary: All Result rows for a ballot — anonymous, short-cached
      description: |
        Served with `Cache-Control: public, max-age=30` so clients can poll
        cheaply while the cron updates provisional tallies every ~10 min.
        Final tallies land the moment `/admin/ballots/{id}/finalize` completes.
      parameters:
        - in: path
          name: ballotId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Array of Result rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Result' }
  /results/proposal/{proposalId}:
    get:
      tags: [Results]
      summary: Single proposal result — anonymous, short-cached
      parameters:
        - in: path
          name: proposalId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Result' }
        '404': { description: No results yet }
...
  /public/results/ballot/{ballotId}:
    get:
      tags: [Public]
      summary: All Result rows for a ballot (provisional + final)
      security: [{ apiKeyHeader: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: ballotId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Array of results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Result' }
  /public/results/proposal/{proposalId}:
    get:
      tags: [Public]
      summary: Single proposal result
      security: [{ apiKeyHeader: [] }, { bearerAuth: [] }]
      parameters:
        - in: path
          name: proposalId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: No results yet }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    Result:
      type: object
      properties:
        proposalId: { type: string }
        ballotId: { type: string, nullable: true }
        source: { type: string, enum: [provisional, final] }
        ballotSource: { type: string, enum: [legacy, hydra] }
        finalizedAt: { type: string, format: date-time, nullable: true }
        # Hydra settlement artifacts (populated only on source=="final")
        hydraEvidenceCid:
          type: string
          nullable: true
          description: IPFS CID of the directory of per-voter evidence files.
        hydraResultsCid:
          type: string
          nullable: true
          description: IPFS CID of the compact results JSON.
        hydraFinalizeTxHash:
          type: string
          nullable: true
          description: Hydra tx hash from the in-head /settle/finalize tx.
        hydraResultsHash:
          type: string
          nullable: true
          description: blake2b_256 of the canonical results JSON — anchored on the (601) datum for on-chain verification.
        hydraEvidenceMerkleRoot:
          type: string
          nullable: true
          description: Merkle root over the per-voter evidence files in the IPFS directory.
        hydraTotalVoters:
          type: integer
          nullable: true
          description: Number of voter tokens included in the final tally.
        hydraExcludedVoters:
          type: array
          description: Voter tokens excluded from the tally (evidence mismatch / missing). Each entry carries tokenName and reason.
          items:
            type: object
            properties:
              tokenName: { type: string }
              reason: { type: string }
        results:
          type: array
          items:
            type: object
            properties:
              id: { oneOf: [{ type: number }, { type: string }] }
              label: { type: string }
              count: { type: integer }
              votingPower: { type: number }
        resultsByGroup: { type: object, nullable: true }